+++ packages/motion-on-scroll/src/helpers/scroll-handler.ts
@@ -8,34 +8,8 @@
 import { play, reverse, setFinalState, setInitialState } from "./animations.js";
 import { DEFAULT_OPTIONS } from "./constants.js";
 import { getPositionIn, getPositionOut, isElementAboveViewport } from "./position-calculator.js";
-import type { ElementOptions } from "./types.js";
+import type { MosElement, ElementOptions } from "./types.js";
 import { debounce, throttle } from "./utils.js";
-
-// ===================================================================
-// TYPES AND INTERFACES
-// ===================================================================
-
-/**
- * Represents an element being tracked for scroll-based animations
- * Contains all state and configuration needed for animation decisions
- */
-export interface MosElement {
-  /** The DOM element being animated */
-  element: HTMLElement;
-  /** Animation configuration options */
-  options: ElementOptions;
-  /** Scroll positions that trigger animations */
-  position: {
-    /** Scroll position where element should animate in */
-    in: number;
-    /** Scroll position where element should animate out (false if disabled) */
-    out: number | false;
-  };
-  /** Whether the element has been animated */
-  animated: boolean;
-  /** Whether the element is currently reversing its animation */
-  isReversing: boolean;
-}
 
 // ===================================================================
 // MODULE STATE
